Rename internal package from `notification` to `pushrules`
